Fix three bugs: --json-response startup crash, wheel build failure, battery-health SQL cast

Verified locally on the PR branch:
- uv build --wheel succeeds (fails on main); all 18 .sql + 18 .toml assets present in the wheel
- teslamate-mcp http --json-response starts and returns plain JSON instead of SSE
- 39 tests pass, ruff check + format clean

+++ src/teslamate_mcp/queries/battery_health_summary.sql
@@ -7,7 +7,7 @@ SELECT c.name as car_name,
     ROUND(
         (
             p.rated_battery_range_km / NULLIF(p.ideal_battery_range_km, 0) * 100
-        ),
+        )::numeric,
         2
     ) as battery_health_pct
 FROM positions p
